+++ src/components/header/header-menu-grid.tsx
@@ -34,7 +34,7 @@ import { Button } from '../shadcn/ui/themed/v2/button'
 import { VScrollPanel } from '../v-scroll-panel'
 
 import { present } from '@/lib/dom-utils'
-import { APP_ICON_CLS, AppIcon } from './app-icon'
+import { APP_ICON_CLS, AppIcon } from '../icons/app-icon'
 import type { IHeaderLinksProps } from './header-menu'
 
 export const SIDE_OVERLAY_CLS = cn(