+++ src/components/icons/favicon.tsx
@@ -7,7 +7,7 @@ import {
 import { cn } from '@/lib/shadcn-utils'
 import { BUTTON_LG_W_CLS } from '@/theme'
 import { VCenterRow } from '../layout/v-center-row'
-import { AppIcon } from './app-icon'
+import { AppIcon } from './edb-icon'
 import { HomeIcon } from './home-icon'
 
 export function FavIcon() {