+++ src/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+- App page icons ui overhaul.
 - Network legend has better reflow for larger fonts with variable spacing as font size increases.
 - Network hidden nodes can now be shown as translucent with adjustable opacity.
 - Network has ability to hide nodes.